change db column

+++ db/migrate/20180404095324_create_results.rb
@@ -3,7 +3,7 @@ class CreateResults < ActiveRecord::Migration[5.2]
     create_table :results do |t|
       t.references :athlete, foreign_key: true
       t.float :score, default: 0.0
-      t.integer :rating_collection, array: true, default: []
+      t.float :rating_collection, array: true, default: []
       t.integer :number_of_votes, default: 0
 
       t.timestamps
